Richard/flask

+++ models/leave.py
@@ -6,6 +6,16 @@ Control Leave processing
 from models.db import DB
 
 
+def is_valid_objectid(id):
+    try:
+        from bson import ObjectId
+        from bson.errors import InvalidId
+        ObjectId(id)
+        return True
+    except (InvalidId, TypeError):
+        return False
+
+
 class LeaveManager:
     """ Leave Manager class """
 
@@ -23,32 +33,41 @@ class LeaveManager:
 
     def get_all_applications_by_user(self, userid):
         try:
-            return self._db.all_by_user(user_id=userid)
+            if not is_valid_objectid(userid):
+                raise ValueError("Application does not exist")
+            return self._db.all_by_user(userid=userid)
         except Exception as e:
             raise e
 
-    def delete_leave_id(self, user, leave_id):
+    def delete_leave(self, user, leave_id):
         try:
+            if not is_valid_objectid(leave_id):
+                raise ValueError("Application does not exist")
             if self._db.delete_application(user.email, leave_id=leave_id):
                 return True
         except ValueError as e:
-            raise ValueError(e)
-        except Exception:
-            raise Exception("An unexpected error occurred while deleting leave application.")
+            raise e
 
     def admin_get_all_applications(self):
         return self._db.all()
 
-    def admin_get_all_user_applications(self, user_id):
-        user = self._db.find_user_by(id=user_id)
-        application = user.lastname, user.firstname, user.email, self.get_all_applications_by_user(user_id)
+    def admin_get_all_user_applications(self, userid):
+        if not is_valid_objectid(userid):
+            raise ValueError("Application does not exist")
+        user = self._db.find_user_by(id=userid)
+        application = (user.lastname, user.firstname, user.email,
+                       self.get_all_applications_by_user(userid))
         return application
 
     def reject_leave_application(self, leave_id):
-        return self._db.approve_reject_leave(leave_id, status=False)
+        if is_valid_objectid(leave_id):
+            return self._db.approve_reject_leave(leave_id, status="rejected")
+        raise ValueError("Application does not exist")
 
     def approve_leave_application(self, leave_id):
-        return self._db.approve_reject_leave(leave_id, status=True)
+        if is_valid_objectid(leave_id):
+            return self._db.approve_reject_leave(leave_id, status="approved")
+        raise ValueError("Application does not exist")
 
     def get_stats(self):
         return self._db.leave_statistics()
